fix: Namespaced PostgreSQL extension did not use namespaced database reference
Added e2e tests to cover this, without the fix, it fails with this
error:

    Warning  CannotResolveResourceReferences  36s (x14 over 77s)  managed/extension.postgresql.sql.m.crossplane.io  cannot resolve references: spec.forProvider.database: cannot get referenced resource: Database.postgresql.sql.m.crossplane.io "example" not found

Fixes #285
