+++ package/crds/postgresql.sql.m.crossplane.io_extensions.yaml
@@ -77,6 +77,9 @@ spec:
                       name:
                         description: Name of the referenced object.
                         type: string
+                      namespace:
+                        description: Namespace of the referenced object
+                        type: string
                       policy:
                         description: Policies for referencing.
                         properties:
@@ -120,6 +123,9 @@ spec:
                         description: MatchLabels ensures an object with matching labels
                           is selected.
                         type: object
+                      namespace:
+                        description: Namespace for the selector
+                        type: string
                       policy:
                         description: Policies for selection.
                         properties:
